Data type changed for total column

+++ Core/ECommerceDomain/Models/Order.cs
@@ -8,7 +8,7 @@ namespace ECommerceDomain.Models
         public int CustomerId { get; set; }
         public string Description { get; set; }
         public string Address { get; set; }
-        public long Total { get; set; }
+        public float Total { get; set; }
 
 
         [ForeignKey("CustomerId")]
